Move internal/x/text to internal/text (#84)
