+++ internal/text/text_test.go
