+++ internal/x/format/base_visitor.go
@@ -27,7 +27,7 @@ import (
 	"text/scanner"
 
 	"github.com/emicklei/proto"
-	"github.com/uber/prototool/internal/x/text"
+	"github.com/uber/prototool/internal/text"
 )
 
 type baseVisitor struct {
