+++ internal/x/format/middle_visitor.go
@@ -25,8 +25,8 @@ import (
 	"strings"
 
 	"github.com/emicklei/proto"
+	"github.com/uber/prototool/internal/text"
 	"github.com/uber/prototool/internal/x/settings"
-	"github.com/uber/prototool/internal/x/text"
 )
 
 type middleVisitor struct {
