+++ internal/x/lint/check_comments_no_c_style.go
@@ -24,7 +24,7 @@ import (
 	"text/scanner"
 
 	"github.com/emicklei/proto"
-	"github.com/uber/prototool/internal/x/text"
+	"github.com/uber/prototool/internal/text"
 )
 
 var commentsNoCStyleChecker = NewAddChecker(
