+++ internal/x/lint/check_enum_field_names_uppercase.go
@@ -23,7 +23,7 @@ package lint
 import (
 	"github.com/emicklei/proto"
 	"github.com/uber/prototool/internal/strs"
-	"github.com/uber/prototool/internal/x/text"
+	"github.com/uber/prototool/internal/text"
 )
 
 var enumFieldNamesUppercaseChecker = NewAddChecker(
