+++ internal/x/lint/check_enum_field_prefixes.go
@@ -25,7 +25,7 @@ import (
 
 	"github.com/emicklei/proto"
 	"github.com/uber/prototool/internal/strs"
-	"github.com/uber/prototool/internal/x/text"
+	"github.com/uber/prototool/internal/text"
 )
 
 var enumFieldPrefixesChecker = NewAddChecker(
