+++ internal/x/lint/check_enum_zero_values_invalid.go
@@ -25,7 +25,7 @@ import (
 
 	"github.com/emicklei/proto"
 	"github.com/uber/prototool/internal/strs"
-	"github.com/uber/prototool/internal/x/text"
+	"github.com/uber/prototool/internal/text"
 )
 
 var enumZeroValuesInvalidChecker = NewAddChecker(
