+++ internal/x/lint/check_file_options_equal.go
@@ -24,7 +24,7 @@ import (
 	"strings"
 
 	"github.com/emicklei/proto"
-	"github.com/uber/prototool/internal/x/text"
+	"github.com/uber/prototool/internal/text"
 )
 
 var fileOptionsEqualGoPackagePbSuffixChecker = NewAddChecker(
