+++ internal/x/lint/check_message_fields_not_floats.go
@@ -22,7 +22,7 @@ package lint
 
 import (
 	"github.com/emicklei/proto"
-	"github.com/uber/prototool/internal/x/text"
+	"github.com/uber/prototool/internal/text"
 )
 
 var messageFieldsNotFloatsChecker = NewAddChecker(
