+++ internal/x/lint/check_package_is_declared.go
@@ -24,7 +24,7 @@ import (
 	"text/scanner"
 
 	"github.com/emicklei/proto"
-	"github.com/uber/prototool/internal/x/text"
+	"github.com/uber/prototool/internal/text"
 )
 
 var packageIsDeclaredChecker = NewAddChecker(
