+++ internal/x/lint/check_package_lower_snake_case.go
@@ -25,7 +25,7 @@ import (
 
 	"github.com/emicklei/proto"
 	"github.com/uber/prototool/internal/strs"
-	"github.com/uber/prototool/internal/x/text"
+	"github.com/uber/prototool/internal/text"
 )
 
 var packageLowerSnakeCaseChecker = NewAddChecker(
