+++ internal/x/lint/check_packages_same_in_dir.go
@@ -26,7 +26,7 @@ import (
 	"text/scanner"
 
 	"github.com/emicklei/proto"
-	"github.com/uber/prototool/internal/x/text"
+	"github.com/uber/prototool/internal/text"
 )
 
 var packagesSameInDirChecker = NewAddChecker(
