+++ internal/x/lint/check_request_response_names_match_rpc.go
@@ -22,7 +22,7 @@ package lint
 
 import (
 	"github.com/emicklei/proto"
-	"github.com/uber/prototool/internal/x/text"
+	"github.com/uber/prototool/internal/text"
 )
 
 var requestResponseNamesMatchRPCChecker = NewAddChecker(
