+++ internal/x/lint/check_request_response_types_in_same_file.go
@@ -25,7 +25,7 @@ import (
 	"text/scanner"
 
 	"github.com/emicklei/proto"
-	"github.com/uber/prototool/internal/x/text"
+	"github.com/uber/prototool/internal/text"
 )
 
 var requestResponseTypesInSameFileChecker = NewAddChecker(
