+++ internal/x/lint/check_request_response_types_unique.go
@@ -22,7 +22,7 @@ package lint
 
 import (
 	"github.com/emicklei/proto"
-	"github.com/uber/prototool/internal/x/text"
+	"github.com/uber/prototool/internal/text"
 )
 
 var requestResponseTypesUniqueChecker = NewAddChecker(
