+++ internal/x/lint/check_rpc_names_camel_case.go
@@ -23,7 +23,7 @@ package lint
 import (
 	"github.com/emicklei/proto"
 	"github.com/uber/prototool/internal/strs"
-	"github.com/uber/prototool/internal/x/text"
+	"github.com/uber/prototool/internal/text"
 )
 
 var rpcNamesCamelCaseChecker = NewAddChecker(
