+++ internal/x/lint/check_services_have_comments.go
@@ -25,7 +25,7 @@ import (
 	"strings"
 
 	"github.com/emicklei/proto"
-	"github.com/uber/prototool/internal/x/text"
+	"github.com/uber/prototool/internal/text"
 )
 
 var servicesHaveCommentsChecker = NewAddChecker(
