+++ internal/x/lint/runner.go
@@ -21,8 +21,8 @@
 package lint
 
 import (
+	"github.com/uber/prototool/internal/text"
 	"github.com/uber/prototool/internal/x/file"
-	"github.com/uber/prototool/internal/x/text"
 	"go.uber.org/zap"
 )
